pyproject.toml
[project]
name = "koo"
version = "0.1.0"
description = ""
authors = [
    {name = "daeun503",email = "zerozero7bang@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "typer (>=0.21.0,<0.22.0)",
    "rich (>=14.2.0,<15.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pymilvus (>=2.6.6,<3.0.0)",
    "openai (>=2.14.0,<3.0.0)",
    "pydantic-ai-slim[openai] (>=1.39.0,<2.0.0)",
    "dependency-injector (>=4.48.3,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "notion-client (>=2.7.0,<3.0.0)",
    "slack-sdk (>=3.39.0,<4.0.0)",
    "alembic (>=1.17.2,<2.0.0)"
]

[tool.poetry]
name = "koo"
version = "0.1.0"
packages = [{ include = "app" }]

[tool.poetry.scripts]
koo = "app.cli:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.10,<0.15.0)"
]

[tool.ruff]
line-length = 120
exclude = ["venv/", ".venv/", "__pypackages__/", "alembic/"]